Further improve error handling
First, we don't want to pass the ERR traps to subshells, because this
will result in the handler run twice. Instead, we make sure errexit is
enabled, so we can catch errors on the outer layer.

Also, we should not directly pipe the result of curl into jq, because on
error, both processes will fail. Instead, we write the result into a
variable, and echo it into jq.